support-swarm/requirements.txt
# Customer Support Swarm - Python Backend
# Dependencies for multi-agent AI support system

# LangChain & Gemini
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# FastAPI & WebSocket
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# RAG - Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
